reworking import logic - removing circular import by introducting cpcommon - reduce hanging imports - adding -i switch for cpr8x for just parsing native config file
