+++ .vscode/launch.json
@@ -100,12 +100,12 @@
                 "PYTHONPATH": "${PYTHONPATH}:${workspaceRoot}"
             },
             "args": [
-                "-m6",
+                "-m11",
                 "-d4",
                 // "-f",
                 //"-c",
                 //  41 - lab fortimanager
-                "-i/home/tim/tmp/configs/mgm_id_56_config_native.json.anon", // all rules, all objs local mgm_id 4
+                //"-i/home/tim/tmp/configs/mgm_id_56_config_native.json.anon", // all rules, all objs local mgm_id 4
                 // "-i/home/tim/tmp/configs/mgm_id_56_config_native.json.anon", // 30k rules, local mgm_id 45
                 //"-i/home/tim/tmp/configs/mgm_id_52_config_native.json.anon", // many gateways, local mgm_id 49
                 //"-i/home/tim/tmp/configs/mgm_id_53_config_native.json.anon", // 1500 rules, 450 NAT rules, local mgm_id 4
@@ -116,6 +116,7 @@
                 //"-i/home/tim/tmp/configs/mgm_id_12_config_native.json",
                 //"-i/home/tim/tmp/configs/mgm_id_53_config_native.json.anon",
                 // "-i/home/tim/tmp/configs/m56_1sthalf.json", // half of 30k rules, local mgm_id 45
+                "-i/home/tim/tmp/configs/mgm_id_1176_config_native.json.anon_error_in_fprs100-s999-rz99-xxx-PolicyNetwork.anon",
                 "-l250"
             ]
         },
@@ -203,6 +204,8 @@
                 "PYTHONPATH": "${PYTHONPATH}:${workspaceRoot}"
             },
             "args": [
+                //"-c",
+                //"-f",
                 "-d0"
             ]
         },
