+++ roles/importer/files/importer/checkpointR8x/parse_rule_csv.py
@@ -1,6 +1,6 @@
 import logging
 import json
-import common, fwcommon, parse_rule
+import common, cpcommon, parse_rule
 
 
 def create_section_header(section_name, layer_name, import_id, rule_uid, rule_num, section_header_uids, parent_uid):
@@ -12,13 +12,13 @@ def create_section_header(section_name, layer_name, import_id, rule_uid, rule_nu
     header_rule_csv += common.csv_add_field('False')            # rule_disabled
     header_rule_csv += common.csv_add_field('False')            # rule_src_neg
     header_rule_csv += common.csv_add_field('Any')              # rule_src
-    header_rule_csv += common.csv_add_field(fwcommon.any_obj_uid) # rule_src_refs
+    header_rule_csv += common.csv_add_field(cpcommon.any_obj_uid) # rule_src_refs
     header_rule_csv += common.csv_add_field('False')            # rule_dst_neg
     header_rule_csv += common.csv_add_field('Any')              # rule_dst
-    header_rule_csv += common.csv_add_field(fwcommon.any_obj_uid) # rule_dst_refs
+    header_rule_csv += common.csv_add_field(cpcommon.any_obj_uid) # rule_dst_refs
     header_rule_csv += common.csv_add_field('False')            # rule_svc_neg
     header_rule_csv += common.csv_add_field('Any')              # rule_svc
-    header_rule_csv += common.csv_add_field(fwcommon.any_obj_uid) # rule_svc_refs
+    header_rule_csv += common.csv_add_field(cpcommon.any_obj_uid) # rule_svc_refs
     header_rule_csv += common.csv_add_field('Accept')           # rule_action
     header_rule_csv += common.csv_add_field('Log')              # rule_track
     header_rule_csv += common.csv_add_field('Policy Targets')   # rule_installon
@@ -83,7 +83,7 @@ def csv_dump_rule(rule, layer_name, import_id, rule_num, parent_uid):
                 elif src['type'] == 'access-role':
                     if isinstance(src['networks'], str):  # just a single source
                         if src['networks'] == 'any':
-                            rule_src_ref += src['uid'] + '@' + fwcommon.any_obj_uid + common.list_delimiter
+                            rule_src_ref += src['uid'] + '@' + cpcommon.any_obj_uid + common.list_delimiter
                         else:
                             rule_src_ref += src['uid'] + '@' + src['networks'] + common.list_delimiter
                     else:  # more than one source
