+++ roles/importer/files/importer/import-main-loop.py
@@ -34,6 +34,10 @@ if __name__ == '__main__':
                         help='proxy server string to use, e.g. http://1.2.3.4:8080')
     parser.add_argument('-s', '--ssl', metavar='ssl_verification_mode', default='',
                         help='[ca]certfile, if value not set, ssl check is off"; default=empty/off')
+    parser.add_argument('-c', '--clear', action='store_true', default=False,
+                    help='If set all imports will run once to delete all data instead of importing')
+    parser.add_argument('-f', '--force', action='store_true', default=False,
+                    help='If set all imports will be run without checking for changes before')
 
     args = parser.parse_args()
     debug_level = int(args.debug)
@@ -125,12 +129,14 @@ if __name__ == '__main__':
                             if not skipping and mgm_details["deviceType"]["id"] in (9, 11):  # only handle CPR8x and fortiManager
                                 logging.debug("import-main-loop: starting import of mgm_id=" + id)
                                 try:
-                                    import_result = common.import_management(mgm_id=id, ssl=args.ssl, debug_level=debug_level, limit=str(api_fetch_limit), proxy=args.proxy)
+                                    import_result = common.import_management(mgm_id=id, ssl=args.ssl, debug_level=debug_level, 
+                                        clearManagementData=args.clear, force=args.force, limit=str(api_fetch_limit), proxy=args.proxy)
                                 except (common.FwoApiFailedLockImport, common.FwLoginFailed):
                                     pass # minor errors for a single mgm, go to next one # logging.debug("Login to firewall failed for mgm_id: " + id)
                                 except: # all other exceptions are logged here
                                     logging.error("import-main-loop - unspecific error while importing mgm_id=" + str(id) + ", " +  str(traceback.format_exc()))
-                                    
+        if args.clear:
+            break # while loop                                    
         if not killer.kill_now:
             logging.info("import-main-loop.py: sleeping between loops for " + str(sleep_timer) + " seconds")
         counter=0
@@ -138,4 +144,4 @@ if __name__ == '__main__':
             time.sleep(1)
             counter += 1
 
-    logging.info("importer-main-loop was killed gracefully.")
+    logging.info("importer-main-loop exited gracefully.")
